requirements.txt
Django==1.11.2
django-crispy-forms==1.6.1
django-formtools==2.0
django-reversion==2.0.8
djangorestframework==3.6.3
future==0.16.0
httplib2==0.9.2
pytz==2017.2
six==1.10.0
xadmin==0.6.1
